bug fix for Last

+++ update.py
@@ -4,21 +4,22 @@ from helpers import *
 
 def get_posts_update_db():
     last = Last.query.first()
-    t = tavorite.userStream(count=200, since_id=last.post_id)['data']
-    for x in t:
-        if x['entities']['links']:
-            if x.get('id') == x.get('thread_id'):
-                if not Post.query.filter_by(post_id=x.get('id')).first():
-                    a = Post(x)
-                    db.session.add(a)
-                    db.session.commit()
+    t = tavorite.userStream(count=200, since_id=last.post_id)
+    if t.get('meta').get('code') == 200:    
+        for x in t['data']:
+            if x['entities']['links']:
+                if x.get('id') == x.get('thread_id'):
+                    if not Post.query.filter_by(post_id=x.get('id')).first():
+                        a = Post(x)
+                        db.session.add(a)
+                        db.session.commit()
     
-    db.session.delete(last)
-    db.session.commit()
+        db.session.delete(last)
+        db.session.commit()
 
-    track_last = Last(t[-1])
-    db.session.add(track_last)
-    db.session.commit()
+        track_last = Last(t[-1])
+        db.session.add(track_last)
+        db.session.commit()
